# docs/OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DBaaS Aggregator API",
    "description": "This documentation presents the REST API for the \"Database as a Service\" (DBaaS) component. DBaaS acts as an aggregator for all adapters. It is designed to collect requests for managed databases and route them to the appropriate adapter. DBaaS stores information about all databases used in a cloud project. These databases are isolated by namespace. DBaaS uses a Classifier to identify databases within a cloud namespace. The Classifier includes service-related information such as scope, microservice name, tenant ID, and namespace.",
    "version": "main"
  },
...
  "paths": {
...
    "/api/v3/dbaas/users/restore": {
      "post": {
        "tags": [
          "Database users controller v3"
        ],
        "summary": "V3. Restore users",
        "description": "The API allows to restore users for database. If role field is passed, then only users with this role will be processed.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreUsersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users recreated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessfulRestoreUsersResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error during restore",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RestoreUsersResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
    "/api/v3/dbaas/users/rotate-password": {
      "post": {
        "tags": [
          "Database users controller v3"
        ],
        "summary": "V3. Rotate user password.",
        "description": "The API allows to rotate password for specific user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserOperationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password was successfully reset.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RestoreUsersRequest": {
        "required": [
          "classifier",
          "type"
        ],
        "type": "object",
        "properties": {
          "classifier": {
            "description": "Classifier describes the purpose of the database and distinguishes this database from other database in the same namespase. It contains such keys as scope, microserviceName, namespace.",
            "type": "object",
            "additionalProperties": {}
          },
          "type": {
            "description": "The physical type of logical database. For example postgresql",
            "type": "string"
          },
          "role": {
            "description": "User role",
            "type": "string"
          }
        }
      },
      "RestoreUsersResponse": {
        "type": "object",
        "properties": {
          "unsuccessfully": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnsuccessfulRestore"
            }
          },
          "successfully": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuccessfullRestore"
            }
          }
        }
      },
...
      "SuccessfulRestoreUsersResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "SuccessfullRestore": {
        "type": "object",
        "properties": {
          "connectionProperties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "UnsuccessfulRestore": {
        "type": "object",
        "properties": {
          "connectionProperties": {
            "type": "object",
            "additionalProperties": {}
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
...
      "UserOperationRequest": {
        "type": "object",
        "properties": {
          "classifier": {
            "description": "Classifier describes the purpose of the database and distinguishes this database from other database in the same namespase. It contains such keys as dbClassifier, scope, microserviceName, namespace. Setting keys depends on the database type.",
            "type": "object",
            "additionalProperties": {}
          },
          "logicalUserId": {
            "description": "User uniq identifier. Using this field with classifier and type dbaas will determine to create or return created before user.",
            "type": "string"
          },
          "type": {
            "description": "The physical type of logical database. For example mongodb or postgresql",
            "type": "string"
          },
          "userId": {
            "description": "The user identificator.",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SecurityScheme": {
        "type": "http",
        "description": "Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}